requirements.txt
aws-cdk-lib==2.189.0
constructs>=10.0.0,<11.0.0
PyYAML==6.0.2
boto3==1.38.7
botocore==1.38.7